Move RestExceptionMapper
