+++ manager/api/micro/src/main/java/io/apiman/manager/api/micro/ManagerApiMicroServiceApplication.java
@@ -16,20 +16,8 @@
 package io.apiman.manager.api.micro;
 
 import io.apiman.manager.api.config.Version;
-import io.apiman.manager.api.rest.impl.ActionResourceImpl;
-import io.apiman.manager.api.rest.impl.ApiManagerApplication;
-import io.apiman.manager.api.rest.impl.CurrentUserResourceImpl;
-import io.apiman.manager.api.rest.impl.DownloadResourceImpl;
-import io.apiman.manager.api.rest.impl.GatewayResourceImpl;
-import io.apiman.manager.api.rest.impl.OrganizationResourceImpl;
-import io.apiman.manager.api.rest.impl.PermissionsResourceImpl;
-import io.apiman.manager.api.rest.impl.PluginResourceImpl;
-import io.apiman.manager.api.rest.impl.PolicyDefinitionResourceImpl;
-import io.apiman.manager.api.rest.impl.RoleResourceImpl;
-import io.apiman.manager.api.rest.impl.SearchResourceImpl;
-import io.apiman.manager.api.rest.impl.SystemResourceImpl;
-import io.apiman.manager.api.rest.impl.UserResourceImpl;
-import io.apiman.manager.api.rest.impl.mappers.RestExceptionMapper;
+import io.apiman.manager.api.rest.impl.*;
+import io.apiman.manager.api.rest.exceptions.mappers.RestExceptionMapper;
 import io.swagger.jaxrs.config.BeanConfig;
 
 import java.util.HashSet;
